+++ tests/middleware/test_exception_handler_middleware.py
@@ -6,10 +6,16 @@ from starlette.exceptions import HTTPException as StarletteHTTPException
 
 from starlite import Request, Response, Starlite, get
 from starlite.config.logging import LoggingConfig
-from starlite.exceptions import HTTPException
+from starlite.exceptions import (
+    HTTPException,
+    InternalServerException,
+    ValidationException,
+)
 from starlite.middleware.exceptions import ExceptionHandlerMiddleware
-from starlite.status_codes import HTTP_500_INTERNAL_SERVER_ERROR
+from starlite.middleware.exceptions.middleware import get_exception_handler
+from starlite.status_codes import HTTP_400_BAD_REQUEST, HTTP_500_INTERNAL_SERVER_ERROR
 from starlite.testing import TestClient, create_test_client
+from starlite.types import ExceptionHandlersMap
 
 if TYPE_CHECKING:
     from _pytest.logging import LogCaptureFixture
@@ -156,3 +162,33 @@ def test_exception_handler_middleware_debug_logging(
         else:
             assert not caplog.records
             assert "exception raised on http connection request to route /test" not in response.text
+
+
+def handler(_: Any, __: Any) -> Any:
+    return None
+
+
+def handler_2(_: Any, __: Any) -> Any:
+    return None
+
+
+@pytest.mark.parametrize(
+    ["mapping", "exc", "expected"],
+    [
+        ({}, Exception, None),
+        ({HTTP_400_BAD_REQUEST: handler}, ValidationException(), handler),
+        ({InternalServerException: handler}, InternalServerException(), handler),
+        ({HTTP_500_INTERNAL_SERVER_ERROR: handler}, Exception(), handler),
+        ({TypeError: handler}, TypeError(), handler),
+        ({Exception: handler}, ValidationException(), handler),
+        ({ValueError: handler}, ValidationException(), handler),
+        ({ValidationException: handler}, Exception(), None),
+        ({HTTP_500_INTERNAL_SERVER_ERROR: handler}, ValidationException(), None),
+        ({HTTP_500_INTERNAL_SERVER_ERROR: handler, HTTPException: handler_2}, ValidationException(), handler_2),
+        ({HTTPException: handler, ValidationException: handler_2}, ValidationException(), handler_2),
+        ({HTTPException: handler, ValidationException: handler_2}, InternalServerException(), handler),
+        ({HTTP_500_INTERNAL_SERVER_ERROR: handler, HTTPException: handler_2}, InternalServerException(), handler),
+    ],
+)
+def test_get_exception_handler(mapping: ExceptionHandlersMap, exc: Exception, expected: Any) -> None:
+    assert get_exception_handler(mapping, exc) == expected
